+++ appyx-interactions/common/src/iosMain/kotlin/com/bumble/appyx/interactions/SystemClock.kt
@@ -1,0 +1,8 @@
+package com.bumble.appyx.interactions
+
+import platform.CoreFoundation.CFAbsoluteTimeGetCurrent
+
+actual object SystemClock {
+
+    actual fun nanoTime(): Long = CFAbsoluteTimeGetCurrent().toLong()
+}