+++ appyx-interactions/common/src/iosMain/kotlin/com/bumble/appyx/interactions/UUID.kt
@@ -1,0 +1,9 @@
+package com.bumble.appyx.interactions
+
+import platform.Foundation.NSUUID
+
+actual object UUID {
+
+    actual fun randomUUID(): String = NSUUID().UUIDString()
+
+}