+++ appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/platform/PlatformBackHandler.kt
@@ -4,6 +4,6 @@ import androidx.compose.runtime.Composable
 
 @Composable
 expect fun PlatformBackHandler(
-    enabled: Boolean = true,
+    enabled: Boolean,
     onBack: () -> Unit
 )