+++ appyx-navigation/common/src/iosMain/kotlin/com/bumble/appyx/navigation/integration/MainIntegrationPoint.kt
@@ -1,0 +1,24 @@
+package com.bumble.appyx.navigation.integration
+
+import com.bumble.appyx.navigation.integrationpoint.IntegrationPoint
+import platform.UIKit.UIViewController
+import platform.UIKit.navigationController
+
+class MainIntegrationPoint: IntegrationPoint() {
+    private lateinit var viewController: UIViewController
+
+    override val isChangingConfigurations: Boolean
+        get() = false
+
+    fun setViewController(viewController: UIViewController) {
+        this.viewController = viewController
+    }
+
+    override fun onRootFinished() {
+        viewController.dismissModalViewControllerAnimated(false)
+    }
+
+    override fun handleUpNavigation() {
+        viewController.navigationController?.popViewControllerAnimated(false)
+    }
+}