+++ demos/appyx-navigation/iosApp/iosApp/ContentView.swift
@@ -1,0 +1,18 @@
+import UIKit
+import SwiftUI
+import ios
+
+struct ComposeView: UIViewControllerRepresentable {
+    func makeUIViewController(context: Context) -> UIViewController {
+        Main_iosKt.MainViewController()
+    }
+
+    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {}
+}
+
+struct ContentView: View {
+    var body: some View {
+        ComposeView()
+                .ignoresSafeArea(.all)
+    }
+}