+++ utils/multiplatform/src/iosMain/kotlin/com/bumble/appyx/utils/multiplatform/Parcelable.kt
@@ -1,0 +1,8 @@
+package com.bumble.appyx.utils.multiplatform
+
+actual annotation class Parcelize
+
+actual interface Parcelable
+
+@Target(AnnotationTarget.TYPE)
+actual annotation class RawValue