+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 ### Added
 
 - [#579](https://github.com/bumble-tech/appyx/pull/579) – Expose `AndroidLifecycle` in `PlatformLifecycleRegistry` for Android
+- [#601](https://github.com/bumble-tech/appyx/pull/601) – Support iOS platform for  **:appyx-navigation** & adding sample iOS targets in **:demos:appyx-navigation** & **:demos:appyx-interactions**  
 
 ### Fixed