+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 ### Added
 
 - [#579](https://github.com/bumble-tech/appyx/pull/579) – Expose `AndroidLifecycle` in `PlatformLifecycleRegistry` for Android
+- [#601](https://github.com/bumble-tech/appyx/pull/601) – iOS support and target apps  
 
 ### Fixed